seperate signature parsing into file
